Fixes for windows + premake updates

+++ Src/Math/FxMathUtil.hpp
@@ -10,8 +10,8 @@
 
 namespace FxMath {
 
-constexpr float32 DegreesToRadians(float32 degrees) { return ((degrees) * (M_PI / 180.0f)); }
-constexpr float32 RadiansToDegrees(float32 radians) { return ((radians) * (180.0f / M_PI)); }
+constexpr float32 DegreesToRadians(float32 degrees) { return ((degrees) * static_cast<float32>(M_PI / 180.0)); }
+constexpr float32 RadiansToDegrees(float32 radians) { return ((radians) * static_cast<float32>(180.0 / M_PI)); }
 
 void SinCos(float32 rad, float32* out_sine, float32* out_cosine);
